pyproject.toml
[project]
name = "open-notes"
version = "0.1.0"
description = "100% local, offline, open-source AI knowledge base with MCP server"
homepage = "https://github.com/Darth-218/open-notes"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "open-notes team" }
]
keywords = ["knowledge-base", "ai", "rag", "mcp", "local", "offline"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "python-frontmatter>=1.1.0",
    "watchdog>=4.0.0",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "llama-cpp-python>=0.2.0",
    "mcp>=1.0.0",
    "click>=8.1.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
open-notes = "open_notes.cli.main:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
